requirements.txt
google-api-python-client==2.115.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests==2.25.1
rich==13.7.0
simple-term-menu==1.6.4